fix: UI Camera space
