+++ Assets/Resources/Prefabs/UI/Instructions.prefab
@@ -48,7 +48,7 @@ Canvas:
   m_GameObject: {fileID: 1944170292936808755}
   m_Enabled: 1
   serializedVersion: 3
-  m_RenderMode: 0
+  m_RenderMode: 1
   m_Camera: {fileID: 3296514539812408186, guid: cdf998b0bea96184d83444008a980ee3, type: 3}
   m_PlaneDistance: 100
   m_PixelPerfect: 0
@@ -284,3 +284,4 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   _instruction: {fileID: 3303263184857114915}
+  _canvas: {fileID: 85677348739216908}
