+++ Assets/Scripts/CameraController.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+
+public class CameraController : MonoBehaviour
+{
+    public Camera GetCamera => GetComponent<Camera>();
+}
